All Tabs but FaceCast working with Layouts
The video in the FaceCast tab is very finicky to work with. It does not maintain it's original aspect ratio and the Video Player 'Aspect Ratio' element seems to do nothing.
All other tabs, including item views, should now be working with layout and look pretty good on all mobile and tablet devices.